scripts/requirements.txt
psycopg2-binary
requests
discord.py
